requirements.txt
streamlit==1.37.1
pandas
openai
ydata-profiling
python-dotenv